+++ src/module/utils/utils.ts
@@ -45,3 +45,23 @@ export function getCharShortName(char: string): string {
       return game.i18n.localize("TWODSIX.Items.Skills." + char);
   }
 }
+
+export function simplifySkillName(skillName:string): string {
+  return skillName.replace(/\W/g, "");
+}
+
+export function ObjectbyString(o, s) {
+  //https://stackoverflow.com/questions/6491463/accessing-nested-javascript-objects-and-arrays-by-string-path
+  s = s.replace(/\[(\w+)\]/g, '.$1'); // convert indexes to properties
+  s = s.replace(/^\./, '');           // strip a leading dot
+  const a = s.split('.');
+  for (let i = 0, n = a.length; i < n; ++i) {
+    const k = a[i];
+    if (k in o) {
+      o = o[k];
+    } else {
+      return;
+    }
+  }
+  return o;
+}